refactor: Mypage - 컴포넌트 분리 & 댓글 수 추가
